pyproject.toml
[project]
name = "django-htmx-complete"
version = "1.0.0"
description = "Complete Django + HTMX application with real-time validation"
authors = [
    {name = "Django HTMX Team", email = "info@djangohtmx.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["django", "htmx", "web", "forms", "validation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Django>=5.1,<5.2",
    "python-decouple>=3.8",
    "Pillow>=10.0.0",
    "django-debug-toolbar>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.2.0",
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "coverage>=7.3.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

production = [
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.7",
    "whitenoise>=6.6.0",
    "sentry-sdk>=1.38.0",
]

[project.urls]
Homepage = "https://github.com/example/django-htmx-complete"
Documentation = "https://github.com/example/django-htmx-complete#readme"
Repository = "https://github.com/example/django-htmx-complete"
"Bug Tracker" = "https://github.com/example/django-htmx-complete/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["accounts", "core", "config"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["accounts", "core", "config"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]